integration tests: move WiFi integration tests to a separate file
