+++ tests/integration.py
@@ -365,27 +365,6 @@ class NetworkTestBase(unittest.TestCase):
 
 class _CommonTests:
 
-    @unittest.skip("Unsupported matching by driver / wifi matching makes this untestable for now")
-    def test_mapping_for_driver(self):
-        self.setup_ap('hw_mode=b\nchannel=1\nssid=fake net', None)
-        with open(self.config, 'w') as f:
-            f.write('''network:
-  renderer: %(r)s
-  wifis:
-    wifi_ifs:
-      match:
-        driver: mac80211_hwsim
-      dhcp4: yes
-      access-points:
-        "fake net": {}
-        decoy: {}''' % {'r': self.backend})
-        self.generate_and_settle()
-        p = subprocess.Popen(['netplan', 'generate', '--mapping', 'mac80211_hwsim'],
-                             stdout=subprocess.PIPE)
-        out = p.communicate()[0]
-        self.assertEquals(p.returncode, 1)
-        self.assertIn(b'mac80211_hwsim', out)
-
     def test_eth_mtu(self):
         self.setup_eth(None)
         self.start_dnsmasq(None, self.dev_e2_ap)
@@ -544,78 +523,6 @@ class _CommonTests:
         self.generate_and_settle()
         self.assert_iface_up(self.dev_e_client, ['inet6 2600:'], ['inet 192.168'])
 
-    def test_wifi_ipv4_open(self):
-        self.setup_ap('hw_mode=b\nchannel=1\nssid=fake net', None)
-
-        with open(self.config, 'w') as f:
-            f.write('''network:
-  renderer: %(r)s
-  wifis:
-    %(wc)s:
-      dhcp4: yes
-      access-points:
-        "fake net": {}
-        decoy: {}''' % {'r': self.backend, 'wc': self.dev_w_client})
-        self.generate_and_settle()
-        # nm-online doesn't wait for wifis, argh
-        if self.backend == 'NetworkManager':
-            self.nm_wait_connected(self.dev_w_client, 60)
-
-        self.assert_iface_up(self.dev_w_client,
-                             ['inet 192.168.5.[0-9]+/24'],
-                             ['master'])
-        self.assertIn(b'default via 192.168.5.1',  # from DHCP
-                      subprocess.check_output(['ip', 'route', 'show', 'dev', self.dev_w_client]))
-        if self.backend == 'NetworkManager':
-            out = subprocess.check_output(['nmcli', 'dev', 'show', self.dev_w_client],
-                                          universal_newlines=True)
-            self.assertRegex(out, 'GENERAL.CONNECTION.*netplan-%s-fake net' % self.dev_w_client)
-            self.assertRegex(out, 'IP4.DNS.*192.168.5.1')
-        else:
-            out = subprocess.check_output(['networkctl', 'status', self.dev_w_client],
-                                          universal_newlines=True)
-            self.assertRegex(out, 'DNS.*192.168.5.1')
-
-    def test_wifi_ipv4_wpa2(self):
-        self.setup_ap('''hw_mode=g
-channel=1
-ssid=fake net
-wpa=1
-wpa_key_mgmt=WPA-PSK
-wpa_pairwise=TKIP
-wpa_passphrase=12345678
-''', None)
-
-        with open(self.config, 'w') as f:
-            f.write('''network:
-  renderer: %(r)s
-  wifis:
-    %(wc)s:
-      dhcp4: yes
-      access-points:
-        "fake net":
-          password: 12345678
-        decoy: {}''' % {'r': self.backend, 'wc': self.dev_w_client})
-        self.generate_and_settle()
-        # nm-online doesn't wait for wifis, argh
-        if self.backend == 'NetworkManager':
-            self.nm_wait_connected(self.dev_w_client, 60)
-
-        self.assert_iface_up(self.dev_w_client,
-                             ['inet 192.168.5.[0-9]+/24'],
-                             ['master'])
-        self.assertIn(b'default via 192.168.5.1',  # from DHCP
-                      subprocess.check_output(['ip', 'route', 'show', 'dev', self.dev_w_client]))
-        if self.backend == 'NetworkManager':
-            out = subprocess.check_output(['nmcli', 'dev', 'show', self.dev_w_client],
-                                          universal_newlines=True)
-            self.assertRegex(out, 'GENERAL.CONNECTION.*netplan-%s-fake net' % self.dev_w_client)
-            self.assertRegex(out, 'IP4.DNS.*192.168.5.1')
-        else:
-            out = subprocess.check_output(['networkctl', 'status', self.dev_w_client],
-                                          universal_newlines=True)
-            self.assertRegex(out, 'DNS.*192.168.5.1')
-
     def test_mix_bridge_on_bond(self):
         self.setup_eth(None)
         self.addCleanup(subprocess.call, ['ip', 'link', 'delete', 'bond0'], stderr=subprocess.DEVNULL)
@@ -754,44 +661,6 @@ class TestNetworkManager(NetworkTestBase, _CommonTests):
     %(e2c)s: {}''' % {'r': self.backend, 'ec': self.dev_e_client, 'e2c': self.dev_e2_client})
         self.generate_and_settle()
         self.assert_iface_up(self.dev_e_client, [], ['inet6 2600:'])
-
-    def test_wifi_ap_open(self):
-        # we use dev_w_client and dev_w_ap in switched roles here, to keep the
-        # existing device blacklisting in NM; i. e. dev_w_client is the
-        # NM-managed AP, and dev_w_ap the manually managed client
-        with open(self.config, 'w') as f:
-            f.write('''network:
-  wifis:
-    renderer: NetworkManager
-    %(wc)s:
-      dhcp4: yes
-      access-points:
-        "fake net":
-          mode: ap''' % {'wc': self.dev_w_client})
-        self.generate_and_settle()
-
-        # nm-online doesn't wait for wifis, argh
-        self.nm_wait_connected(self.dev_w_client, 60)
-
-        out = subprocess.check_output(['iw', 'dev', self.dev_w_client, 'info'],
-                                      universal_newlines=True)
-        self.assertIn('type AP', out)
-        self.assertIn('ssid fake net', out)
-
-        # connect the other end
-        subprocess.check_call(['ip', 'link', 'set', self.dev_w_ap, 'up'])
-        subprocess.check_call(['iw', 'dev', self.dev_w_ap, 'connect', 'fake net'])
-        out = subprocess.check_output(['dhclient', '-1', '-v', self.dev_w_ap],
-                                      stderr=subprocess.STDOUT, universal_newlines=True)
-        self.assertIn('DHCPACK', out)
-        out = subprocess.check_output(['iw', 'dev', self.dev_w_ap, 'info'],
-                                      universal_newlines=True)
-        self.assertIn('type managed', out)
-        self.assertIn('ssid fake net', out)
-        out = subprocess.check_output(['ip', 'a', 'show', self.dev_w_ap],
-                                      universal_newlines=True)
-        self.assertIn('state UP', out)
-        self.assertIn('inet 10.', out)
 
 
 unittest.main(testRunner=unittest.TextTestRunner(stream=sys.stdout, verbosity=2))
